create reweightmc folder
